2024103432王鑫/项目作业/requirements.txt
geopandas==1.0.1
numpy==1.26
rasterio==1.4.2
scipy==1.13.0
tensorflow==2.18.0
ultralytics
pillow
pyyaml
opencv-python
tqdm
